Enable the GameStateSamples Gem.
This namely provides a simple main menu from which the user can open any level within the project, and a simple pause menu from which the user can return to the main menu in order to select a different level to open.

The user can interact with both menus using mouse/keyboard, gamepad, or touch screen, and the menus are only show when running the game launcher (not the editor).

Signed-off-by: bosnichd <bosnichd@amazon.com>